pyproject.toml
[project]
name = "ai_domain_agent"
version = "0.1.0"
description = "ai-domain-agent using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.141.0,<1.0.0",
    "google-generativeai>=0.8.5",
    "python-whois",
]

[project.scripts]
ai_domain_agent = "ai_domain_agent.main:run"
run_crew = "ai_domain_agent.main:run"
train = "ai_domain_agent.main:train"
replay = "ai_domain_agent.main:replay"
test = "ai_domain_agent.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
